Removes unnecessary auth
