--- services/api-server/src/simcore_service_api_server/api/routes/authentication/api_key.py
@@ -1,15 +1,0 @@
-# FIXME: Until tests
-# pylint: skip-file
-#
-
-from fastapi import APIRouter
-
-router = APIRouter()
-
-
-@router.post("/login", response_model=UserInResponse, name="auth:login")
-async def login(
-    user_login: UserInLogin = Body(..., embed=True, alias="user"),
-    users_repo: UsersRepository = Depends(get_repository(UsersRepository)),
-) -> UserInResponse:
-    pass
